LASB-1832 Additional functionality and placeholders and package changes
