+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/builder/EformsApplicationMapper.java
@@ -1,0 +1,8 @@
+package gov.uk.courtdata.eform.builder;
+
+import gov.uk.courtdata.dto.EformsStagingDTO;
+import gov.uk.courtdata.eform.model.EformsApplication;
+
+public interface EformsApplicationMapper {
+    EformsStagingDTO map(EformsApplication eformsApplication);
+}
