+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/mapper/EformsStagingDTOMapper.java
@@ -1,0 +1,14 @@
+package gov.uk.courtdata.eform.mapper;
+
+import gov.uk.courtdata.dto.EformsStagingDTO;
+import gov.uk.courtdata.entity.EformsStagingEntity;
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring")
+public interface EformsStagingDTOMapper {
+
+    EformsStagingEntity toEformsStagingEntity(EformsStagingDTO eformsStagingDTO);
+
+    EformsStagingDTO toEformsStagingDTO(EformsStagingEntity eformsStagingEntity);
+}
