+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/model/ClientDetails.java
@@ -1,0 +1,14 @@
+package gov.uk.courtdata.eform.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ClientDetails {
+    private Applicant applicant;
+}
