+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/model/Defendant.java
@@ -1,0 +1,23 @@
+package gov.uk.courtdata.eform.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Defendant {
+    @JsonProperty("first_name")
+    private String firstName;
+
+    @JsonProperty("last_name")
+    private String lastName;
+
+    @JsonProperty("conflict_of_interest")
+    private String conflictOfInterest;
+
+}
