+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/model/FromDateToDate.java
@@ -1,0 +1,25 @@
+package gov.uk.courtdata.eform.model;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class FromDateToDate {
+
+    @JsonProperty("date_from")
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+    private LocalDate dateFrom;
+
+    @JsonProperty("date_to")
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+    private LocalDate dateTo;
+}
