+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/model/InterestOfJustice.java
@@ -1,0 +1,16 @@
+package gov.uk.courtdata.eform.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class InterestOfJustice {
+
+    private String type;
+    private String reason;
+}
