+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/model/Offence.java
@@ -1,0 +1,22 @@
+package gov.uk.courtdata.eform.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Offence {
+    private String name;
+
+    @JsonProperty("offence_class")
+    private String offenceClass;
+
+    private List<FromDateToDate> dates;
+}
