+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/repository/EformsStagingRepository.java
@@ -1,0 +1,11 @@
+package gov.uk.courtdata.eform.repository;
+
+import gov.uk.courtdata.entity.EformsStagingEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EformsStagingRepository extends JpaRepository<EformsStagingEntity,Integer> {
+
+
+}
