+++ maat-court-data-api/src/main/java/gov/uk/courtdata/eform/validator/EformsApplicationUsnValidator.java
@@ -1,0 +1,12 @@
+package gov.uk.courtdata.eform.validator;
+
+import gov.uk.courtdata.eform.model.EformsApplication;
+import org.springframework.stereotype.Component;
+
+@Component
+public class EformsApplicationUsnValidator {
+
+    public void validate(EformsApplication eformsApplication, Integer usn) {
+        // TODO validate non-null usn and matching or throw exception
+    }
+}
