--- maat-court-data-api/src/main/java/gov/uk/courtdata/eforms/mapper/EformsStagingDTOMapper.java
@@ -1,14 +1,0 @@
-package gov.uk.courtdata.eforms.mapper;
-
-import gov.uk.courtdata.dto.EformsStagingDTO;
-import gov.uk.courtdata.entity.EformsStagingEntity;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring")
-public interface EformsStagingDTOMapper {
-
-    EformsStagingEntity toEformsStagingEntity(EformsStagingDTO eformsStagingDTO);
-
-    EformsStagingDTO toEformsStagingDTO(EformsStagingEntity eformsStagingEntity);
-}
