--- maat-court-data-api/src/main/java/gov/uk/courtdata/model/eforms/InterestOfJustice.java
@@ -1,16 +1,0 @@
-package gov.uk.courtdata.model.eforms;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class InterestOfJustice {
-
-    private String type;
-    private String reason;
-}
