--- maat-court-data-api/src/main/java/gov/uk/courtdata/model/eforms/xmlModels/FormData.java
@@ -1,18 +1,0 @@
-package gov.uk.courtdata.model.eforms.xmlModels;
-
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-
-public class FormData {
-
-   @JacksonXmlProperty(localName = "fielddata")
-    private FieldData fieldData;
-}
