--- maat-court-data-api/src/main/java/gov/uk/courtdata/model/eforms/xmlModels/LaaAdded.java
@@ -1,19 +1,0 @@
-package gov.uk.courtdata.model.eforms.xmlModels;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.xml.bind.annotation.XmlRootElement;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LaaAdded {
-    @JsonProperty("CaseType")
-    private String caseType;
-}
