--- maat-court-data-api/src/main/java/gov/uk/courtdata/repository/EformsStagingRepository.java
@@ -1,11 +1,0 @@
-package gov.uk.courtdata.repository;
-
-import gov.uk.courtdata.entity.EformsStagingEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EformsStagingRepository extends JpaRepository<EformsStagingEntity,Integer> {
-
-
-}
