+++ maat-court-data-api/src/test/java/gov/uk/courtdata/eform/controller/EformsStagingRestControllerTest.java
@@ -1,0 +1,30 @@
+package gov.uk.courtdata.eform.controller;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class EformsStagingRestControllerTest {
+
+    @BeforeEach
+    void setUp() {
+        // TODO Complete this
+    }
+
+    @Test
+    void updateEformApplication() {
+    }
+
+    @Test
+    void getEformsApplication() {
+    }
+
+    @Test
+    void deleteEformsApplication() {
+    }
+
+    @Test
+    void createEformsApplication() {
+    }
+}
